getXModified now includes all Classes bonus which are added to the score,
Added Enums for Class, Race, SubClass (One for each class) and Ability,
   Race now points to this enum.
Added 2 fields for HalfElf additional 2 abilities,
Ability base fields is now "Score" eg StrengthScore as this is what it is referred to in a Player's Guide,
RaceHelper now redirects itself referencing the Enum on Sheet. It will have to be decided where these Enums are stored.