Stabilised project it now runs.
Sheet:
Weapon list now added to sheet - will now store a list of user's weapon (also added getWeaponList
setModifier is now renamed to getModifier
@Contract has been commented out - no errors have occurred just yet
getAbility now added for Weapon. In weapon it is used for the type of damage it deals. So it also has Finesse which is not used otherwise and checks which is the higher value of STR and DEX

SheetEnum:
All Enums now throw an exception when they are not able to find an the key instead of returning NULL.

Weapon & Spell:
RangeFeet is now RangeNormal - to coincidence with WeaponRangeNormal.
//* is to indicate that Weapon & Spell have a shared property

Weapon:
primaryKey weaponID added
weaponStatBonus changed to an int as it now stores an EnumKey

CombatFragment:
OnCreate retrieves the weaponList
Added DeleteSheetWeapon function - removes weapon which button it is linked too

attack_view_list:
Now horizontal
Removed GridView. it is now a listView item

attack_view_grid_title:
Added may not be used

content_combat:
GridLayout changed to ListView
Title Grid removed will need to be re-added correctly later

strings:
added string-array for damage_type spinner

UnitTestWeapon:
Disabled as it's popping up warning messages on Weapon.java stating that properties have been set when it's only for a test. When I find out how to do it correctly it will be readded

build.gradle:
buildToolsVersion now 26.0.2 was 25.0.3
gradle now 3.0.0 was 2.3.3