Weapon List started.
Weapons.csv asset added (This contains a list of all the weapons the user could import)
WeaponEnum added - has DamageDieType and DamageType within - For the time being I will not stick to the standard in the interest of time, for the time being

weapon:
numberOfDie changed back to int.
weaponStatBonus renamed to weaponAbilityBonus
weaponAbilityBonus linked to ENUM ability - May need additional editing to convert when pulling in data.
WeaponDamageDieType & WeaponDamageType linked to Weapon Enum

AttackListView:
Added NumberPicker

CSVHelper commented. I was using it as help for the CSV Function in MainActivity. Will likely be deleted.

MainActivity:
csvReader function added
This will read in a list of Weapons which the user would be able to pick from if they wish