Added new features to Misc, and solved the issue of the Weapon dialogue box
Sheet
Equipment is added as a List & String Notes is added (Getters and Setters added for both)
getMethod for BonusLangauge added
getRaceLanguage changed to public (oops)
getRaceLanguageAsReadableString added - changes from Language Enum and returns a String (Which the first letter in uppercase)
getRaceName removed (I don't believe this is need as enum exists

MiscFragment & content_misc
changed over ListView of languages to a textView and a spinner)
added - this includes saving/reading to/from sheet, added to the XML.
inspiration, s_bonus_language, notes (I decided to add additional space for notes as there was a lot of unused space).
added - these don't require saving but are also added to the XML.
knownLanguage - This checks the race and gets the language spoken from sheet
bonus_language_title - This is added because both this and s_bonus_language are invisible unless the user has a race which has a bonus language. This is done in visibilityCheck
ERROR - currently the spinner does not save the value, it will be fixed by the next commit but doesn't cause any major issues.

custom_dialogue_list
has had a relatively substation overhaul. It turns out OnClick will NOT respond if there are too many lines (it seems to be a maximum of one use of layout_below so 2 rows, why this is the case it is currently unclear. This was discovered while being suggested to take a closer look and break down things to basics. (Comment was since removed) https://stackoverflow.com/questions/48159144/setonitemclicklistener-in-dialog-does-not-respond-to-press
This lead to placing everything onto two rows, with expanding custom_dialogue to encompass it.
IMPORTANT NOTE
CombatFragment has not been committed as it has code which is being worked on. While the dialogue should work there is every possibility it may not.

Minor changes
removed TODOs which have been completed. TODOs which are not possible to complete have been marked as "UNABLE"

custom_dialogue
width is now set at 1000dp (Which should widen the dialogue box fitting all the information in

Dimen
all fragment vertical separator now have an assigned dimen (v_separator_fragment)
(content - combat, adventure, misc)

strings
added values for
inspiration, bonus_language, notes
as well as language array.

Equipment
id changed from String to Int - to work with checking the highest value.
getters and setters added for ID

MainActivity
TODOs UNABLEs added (Things that were planned but also have noted down)

content_adventure
removed random TODO but not certain as to why it was there. (Calculated  +id/et_perception + 10)
removed saving throws todo, UNABLE(ed) a couple more TODOs here
changed all alignRight to alignEnd

content_misc
removed TODO for inspiration as it was added